requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography>=42